Allow image decoding on smaller buffers and chunked streams
Currently pfim will fail to decode a dds image if there is not enough
data available to decode an entire row of the image. The default buffer
size will be insufficient on large images (like an image with a width of
16384 pixels). This PR fixes the problem by only requiring the buffer to
be big enough to decode a block of pixels instead an entire row.

Pfim will also fail to properly incorporate chunked streams into the
decode process. A chunked stream is one where a `Read` will yield data
but won't fill the buffer in its entirety (and the stream is not
finished). [This is documented behavior][0].

> An implementation is free to return fewer bytes than requested even if
> the end of the stream has not been reached.

The fix is to convert all stream reads so that they fill the buffer provided.

[0]: https://docs.microsoft.com/en-us/dotnet/api/system.io.stream.read?view=net-6.0#definition